move simulation roundtrip to integration tests
